detector/requirements.txt
PyYAML>=6.0
Flask>=3.0
requests>=2.31
